UnknownServiceException an inner class of ServicesMonitor
