--- Core/src/org/sleuthkit/autopsy/core/UnknownServiceException.java
@@ -1,33 +1,0 @@
-/*
- * Autopsy Forensic Browser
- * 
- * Copyright 2012 Basis Technology Corp.
- * Contact: carrier <at> sleuthkit <dot> org
- * 
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *     http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.sleuthkit.autopsy.core;
-
-/**
- * Exception thrown when service status query results in an error.
- */
-public class UnknownServiceException extends Exception {
-
-    public UnknownServiceException(String message) {
-        super(message);
-    }
-
-    public UnknownServiceException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
